hubble: Update the reason label for hubble_drop_total metric
Use DropReasonDesc as the reason label instead of using the deprecated
DropReason field and then converting it using the monitor API. Use of
the monitor API is confusing in one specific case where the drop reason
is unknown, as the reason label gets set to "Success". With this change,
the reason label is set to "DROP_REASON_UNKNOWN" in case the drop reason
is unknown.

Ref: #22402
Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>