test/l4lb: Use DinD instead of Kind/Helm
The usage of Kind/Helm when testing the Cilium in the standalone L4LB
mode started to be a source of confusion. In particular, this gave a
wrong impression that in that mode Cilium still has a connectivity to
the kube-apiserver. This is not true.

Previously, we used Kind/Helm for these tests just to create Docker
containers acting as nodes, and then to install Cilium. This can be
achieved by simply using docker/dind (aka Docker-in-Docker), and then
starting the L4LB Cilium by execing into the dind container and running
"docker run".

Signed-off-by: Martynas Pumputis <m@lambda.lt>